Calculate back edge data in ms not s, replace any made-up data with None
For data that isn't real but could easily be calculated (example:
gap between first frame and... the float 0.0, which is just a
placeholder value) we should make sure to set that data as None,
so later code will error out if we try to do math on it.

The code should refuse to do math on made-up or placeholder data,
rather than allowing made-up data to influence the final stats.