Composite some fake filler frames ("seed frames") to start, if necessary
For captures where the first present time is slower than
the OBS composite interval.

Without this correction, we would composite the first real pframe
some extra times, and it would distort the stats at least a tiny bit.

This catches composite times up to the first real present time,
but doesn't include the fake "filler" composited frames
in the final composited frame lists.

As a side effect, oframes in the output may start from e.g. oframe #2
instead of from oframe #0, since oframes #0 and #1 would be fake filler
frames in this example.

Preserves fidelity of the stats, at the cost of being a bit unintuitive.