Don't calculate or show stddev of a deviations list
Whether deviations have a tight grouping isn't a relevant outcome
to care about.

It could be a tight grouping around 1ms, or a tight grouping around 5ms,
or a loose scatter of data. None of the above would matter.

What matters is magnitude or "size" -- deviations, when tracked,
are hoped to simply be as small (low magnitude) as possible.

Low deviation means high consistency, high consistency is good.

(To put it in a roundabout and not at all clarifying way:
"A tight grouping around 0" is best, coincidentally,
but that's because of the low magnitude (as close to 0 as possible),
not because of the grouping. A loose distribution close to 0 is also
good. Because it means the individual deviations are low.)

tl;dr big deviation bad, small deviation good, grouping incidental
and not important.

Might be *interesting* to calculate, but in fact I have found it
not to be interesting, but instead only serves to be quite distracting.

So... deleting the stddev reports for deviationlists!