Don't do "unique output/composited" stats if redundant to regular "output/composited"
Don't calculate separate "unique output/composited" frame stats
if the list of "unique output/composited" frames is the same
list of frames as the list of all "output/composited" frames.

That is, if there were no doubled/duplicated frames. (No pframes
composited more than once, and so appearing more than once in the
hypothetical stream/recording).

The output/composited frame stats are already massive on this branch.
No-one wants to scan through visually to see this massive list of stats
only to find out that nearly half of them are redundant (exagerrating,
but not by a lot).

Trim these redundant stats out (and finish the sim faster!) if
"unique" is redundant to "all" output/composited frames.